Handle state of fishes object in App
The initial state is set in the constructor - fishes and order are initialized to empty objects. An AddFish method is added and binded to the App component. In this method, we update the state of fishes by making a copy of the existing fishes state, then adding the new fish to the copy and using setState to set to the modified copy. This is React best practices - it increases performance and protects against race conditions. Finally, we 'float' the new fish object from AddFishForm up to App through props.