'Load Sample Fish' functionality
Added a button to load the sample fish file into App state. While the button lives in the Inventory component, state lives at the top level - in the App component. So we build the loadSamples() function in the App component and bind it, then pass it down to the Inventory component via props, so that we can trigger it onClick()