Updating Order State
Created and bound addToOrder function to the App component. addToOrder updates the state of order by incrementing a certain type of fish, given by the parameter key. We propogate addToOrder down to the Fish component through props, and also create a new props for Fish called index to mirror key (key is special, only be used by React). Now in the Fish component, we want to call addToOrder on the click event for the button. We can't do this the normal way due to the function argument (since JSX is rendered as JavaScript, the parentheses on the function will cause it to be invoked onload). We can work around this issue by using a callback and an ES6 arrow function...