Set up the Order Component
Send fishes and order down from App into Order component as props. Return some JSX t
o handle the order screen and proper format.. The renderOrder function was just creat
ed to outsource some work from the render function of this component, and help keep t
he code more organized. In order to use this in the context of this new function, we
do need to add a constructor for the class (dont forget to call super!) and then bind
 'this' to the renderOrder function. The JSX declared in this component should render
 out the amount of each fish ordered, and the total price of the order.