Created Firebase connection, synced FB state to fishes
File base.js uses module 're-base' to keep a connection to our FireBase instance, and we can export that class we created and import it in App.js to immediately work with the FireBase API. In App.js, we are able to sync to FireBase with only a few extra lines of code. We use React Lifecycle Methods to sync state of a particular store to firebase.