Ability to update Fish in Inventory & remove Orders
The fish in the inventory are formatted similar to the AddFishForm component. The user can click on properties of a fish in the inventory and edit them - changes are seen instantaneously in the Fish Menu and Orders areas. Fish can be removed from the inventory. Also added ability to remove Orders. There are two states in the React application - fishes and orders. fishes are stored on FireBase relevant to the store, and orders are stored on the user's localstorage