Couple of Animations using CssTransitionGroup for React
CSSTransitionGroup is not a core part of React, but it is still supported well. Two animations were added to the Order component, by using the CSSTransitionGroup tag in place of the old tag. The old tag's element is passed to the 'component' attribute of CSSTG ex component=ul. Some other tags are relevant too, defining the duration of the transition (aka how long the transition CLASSes are applied) and what the classes will be named. These temporary transition classes are created by React/CSSTG and allow us to target them w/ CSS and apply transformations there